+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Implement `Extend` for `TiVec`.
+
 ### Changed
 - Set `impl-index-from` feature always enabled because of
   compatibility issues between crates that use it and crates that don't.